Fix #1
Use urllib.request.urlopen instead of urllib.request.urlretrieve
to download from a given URL, and if save_dir is given, the downloaded
file is saved directly in the desired directory. Before these changes,
the file was first downloaded to the current working directory and then
moved to the chosen save_dir.